+++ src/components/dialog/dialog.ts
@@ -32,9 +32,10 @@ const hasPreventScroll = isPreventScrollSupported();
  * @event sl-after-hide - Emitted after the dialog closes and all animations are complete.
  * @event sl-initial-focus - Emitted when the dialog opens and the panel gains focus. Calling `event.preventDefault()`
  *   will prevent focus and allow you to set it on a different element in the dialog, such as an input or button.
- * @event sl-request-close - Emitted when the user attempts to close the dialog by clicking the close button, clicking the
- *   overlay, or pressing the escape key. Calling `event.preventDefault()` will prevent the dialog from closing. Avoid
- *   using this unless closing the dialog will result in destructive behavior such as data loss.
+ * @event {{ source: 'close-button' | 'keyboard' | 'overlay' }} sl-request-close - Emitted when the user attempts to
+ *   close the dialog by clicking the close button, clicking the overlay, or pressing escape. Calling
+ *   `event.preventDefault()` will keep the dialog open. Avoid using this unless closing the dialog will result in
+ *   destructive behavior such as data loss.
  *
  * @csspart base - The component's base wrapper.
  * @csspart overlay - The overlay.
@@ -123,8 +124,12 @@ export default class SlDialog extends LitElement {
     return waitForEvent(this, 'sl-after-hide');
   }
 
-  private requestClose() {
-    const slRequestClose = emit(this, 'sl-request-close', { cancelable: true });
+  private requestClose(source: 'close-button' | 'keyboard' | 'overlay') {
+    const slRequestClose = emit(this, 'sl-request-close', {
+      cancelable: true,
+      detail: { source }
+    });
+
     if (slRequestClose.defaultPrevented) {
       const animation = getAnimation(this, 'dialog.denyClose');
       animateTo(this.panel, animation.keyframes, animation.options);
@@ -137,7 +142,7 @@ export default class SlDialog extends LitElement {
   handleKeyDown(event: KeyboardEvent) {
     if (event.key === 'Escape') {
       event.stopPropagation();
-      this.requestClose();
+      this.requestClose('keyboard');
     }
   }
 
@@ -217,7 +222,7 @@ export default class SlDialog extends LitElement {
         })}
         @keydown=${this.handleKeyDown}
       >
-        <div part="overlay" class="dialog__overlay" @click=${this.requestClose} tabindex="-1"></div>
+        <div part="overlay" class="dialog__overlay" @click=${() => this.requestClose('overlay')} tabindex="-1"></div>
 
         <div
           part="panel"
@@ -241,7 +246,7 @@ export default class SlDialog extends LitElement {
                     name="x"
                     label=${this.localize.term('close')}
                     library="system"
-                    @click="${this.requestClose}"
+                    @click="${() => this.requestClose('close-button')}"
                   ></sl-icon-button>
                 </header>
               `